% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aboptim.R
\name{aboptim}
\alias{aboptim}
\title{Find Optimum L2 Regularisation Parameter Across Specified Range}
\usage{
aboptim(ratings_matrix, range_min = 0, range_max = 1, rank = 1)
}
\arguments{
\item{ratings_matrix}{The sparse ratings matrix, containing NAs where no
rating exists, which is to be decomposed.}

\item{range_min}{Lower end of range of values over which to optimise. Default is 0. Must be positive numeric.}

\item{range_max}{Upper end of range of values over which to optimise. Default is 1. Must be positive numeric.}

\item{rank}{Matrix decomposition rank, k, which should be found prior to running aboptim by using
koptim.}
}
\value{
A single parameter for L2 regularisation which the user will then set alpha and beta to
when running nnfm.
}
\description{
Find Optimum L2 Regularisation Parameter Across Specified Range
}
\examples{

Find the optimum L2 regularisation parameter from a range between 0.2 and 1.2

k <- koptim(ratings_matrix = wide_format, 0.2, 1.2)
}
