% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/koptim.R
\name{koptim}
\alias{koptim}
\title{Find Optimum Matrix Decomposition Rank Across Specified Range}
\usage{
koptim(ratings_matrix, k_min = 1, k_max = 5)
}
\arguments{
\item{ratings_matrix}{The sparse ratings matrix, containing NAs where no
rating exists, which is to be decomposed.}

\item{k_min}{Lower end of range of values over which to optimise the rank.
Type is positive integer. Default value is 1.}

\item{k_max}{Upper end of range of values over which to optimise the rank.
Type is positive integer. Default value is 5.}
}
\value{
The rank (ie value of k), for which the given sparse ratings matrix has the lowest MSE
over the given range.
}
\description{
Find Optimum Matrix Decomposition Rank Across Specified Range
}
\examples{
# Find the optimum rank (k value) for a range of k between 1 and 5

k <- koptim(wide_format, 1, 5)

}
