% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixify.R
\name{matrixify}
\alias{matrixify}
\title{Generate Ratings Matrix for Non-Zero Book Ratings in Ratings Data Frame}
\usage{
matrixify(ratings, zeroes = TRUE)
}
\arguments{
\item{ratings}{A table of numeric ratings of type data.frame, with columns: User.ID, ISBN, Book.Rating.}

\item{zeroes}{Boolean - default is TRUE - designating whether missing ratings should be zeroes (TRUE)
or NAs (FALSE). For matrix multiplication, missing values must be NAs.}
}
\value{
An m x n matrix with users as rows and books as columns.
}
\description{
Generate Ratings Matrix for Non-Zero Book Ratings in Ratings Data Frame
}
\examples{
# Create a matrix to be used for user-based collaborative filtering

book_ratings <- data.frame(User.ID = c(277042, 276925),
ISBN = c(0316666343, 0385504209), Book.Rating = c(5, 7))

wide_matrix <- matrixify (ratings = book_ratings, zeroes = TRUE)

}
