% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ub.R
\name{ub}
\alias{ub}
\title{Generate Book Recommendations Using User-Based Collaborative Filtering}
\usage{
ub(new_user = FALSE, reader_id, books, ratings_matrix, new_user_df)
}
\arguments{
\item{new_user}{Boolean (default value is FALSE), designating whether recommendations
for a new or existing user will be made. If set to TRUE, a data frame containing information
about the new user should be passed to the new_user_df argument.}

\item{reader_id}{The numeric ID of the user you wish to generate recommendations for.}

\item{books}{A table of books of type data.frame, with columns: ISBN, Book.Title, Book.Author.}

\item{ratings_matrix}{An m x n matrix containing users on the rows and books on the columns.}

\item{new_user_df}{A data frame containing the information of a new user's reading history.
See examples for usage.}
}
\value{
A data frame containing the top three books (title and author) recommended
 for the user specified.
}
\description{
Generate Book Recommendations Using User-Based Collaborative Filtering
}
\examples{
# Generate recommendations for an existing user:

ub (reader_id = "277042", books = book_info, ratings_matrix = matrixify_ratings)

# Generate recommendations for a new user:

new_reader <- data.frame(ISBN = c('0440234743', '0971880107', '0345417623'),
Book.Rating = c(2, 5, 3))

ub (new_user = TRUE, books = book_info, ratings_matrix = matrixify_ratings,
new_user_df = new_reader)

}
